refactor/split-long-code-file-into-small-file
1. Modular architecture: Split the 920-line monolithic main.js into 7 logical modules:
    - config.js - Configuration constants and DOM element references
    - state.js - Application state management (using default export)
    - textProcessor.js - Text processing functions (splitText, is_content)
    - apiClient.js - API calls and queue management
    - uiRenderer.js - UI rendering and formatting (includes updateApiStatusUI)
    - fileHandlers.js - File handling functionality
    - previewMode.js - Preview mode functionality
  2. Main entry file: Create a new main.js as the application entry point, importing all modules and initializing the application.
  3. HTML update: Modify the script tag in index.html to type="module" to support ES6 modules.
  4. Issue fixes:
    - Resolved circular dependency issue with the updateApiStatusUI function (moved to uiRenderer.js)
    - Updated import relationships between all modules
    - Removed redundant bindRegenerateButtons function (functionality handled by event delegation in the main module)

  File structure

  static/js/
  ├── main.js                 # New entry file (75 lines)
  └── modules/
      ├── config.js          # Configuration (47 lines)
      ├── state.js           # State (29 lines)
      ├── textProcessor.js   # Text processing (~100 lines)
      ├── apiClient.js       # API client (203 lines)
      ├── uiRenderer.js      # UI rendering (337 lines)
      ├── fileHandlers.js    # File handling (~80 lines)
      └── previewMode.js     # Preview mode (143 lines)

  All modules are properly connected, and application functionality remains unchanged. The code structure is clearer and easier to maintain and test.
